[FIX] 【需求35455】【文件管理器】在根目录下搜索时，搜索过程中进度条填满整个搜索框而不是实时显示进度条  [LINK]278359

+++ libpeony-qt/controls/tool-bar/search-bar-container.h
@@ -52,7 +52,7 @@ protected:
 private:
     QVariantAnimation *m_animation = nullptr;
     bool m_searching = false;
-    int m_value = 0;
+    double m_value = 0;
 };
 
 class SearchBarContainer : public QWidget
